+++ ca/django_ca/management/base.py
@@ -151,14 +151,9 @@ class BaseCommand(mixins.ArgumentsMixin, _BaseCommand, metaclass=abc.ABCMeta):
         self, parser: ActionsContainer, default: Optional[str] = "RSA", default_text: str = "%(default)s"
     ) -> None:
         """Add --key-type option (type of private key - RSA/DSA/EC/Ed25519/Ed448)."""
-        # NOTE: This can be simplified once support for "ECC" and "EdDSA" values is dropped.
-        known_private_key_types = ["RSA", "DSA", "EC", "Ed25519", "Ed448"]  # pragma: only django-ca<1.26
-        metavar = f"{{{','.join(known_private_key_types)}}}"
-
         parser.add_argument(
             "--key-type",
-            choices=known_private_key_types + ["ECC", "EdDSA"],
-            metavar=metavar,
+            choices=["RSA", "DSA", "EC", "Ed25519", "Ed448"],
             default=default,
             help=f"Key type for the private key (default: {default_text}).",
         )