+++ ca/django_ca/tests/test_deprecation.py
@@ -21,8 +21,9 @@ from django.test import TestCase
 
 from django_ca.deprecation import (
     DeprecationWarningType,
-    RemovedInDjangoCA126Warning,
     RemovedInDjangoCA127Warning,
+    RemovedInDjangoCA128Warning,
+    RemovedInDjangoCA129Warning,
     RemovedInNextVersionWarning,
     deprecate_argument,
 )
@@ -31,17 +32,22 @@ from django_ca.deprecation import (
 class TestDjangoCATestCase(TestCase):
     """Test :py:mod:`django_ca.deprecation`."""
 
-    msg_in_126 = "deprecated in 1.26"
     msg_in_127 = "deprecated in 1.27"
+    msg_in_128 = "deprecated in 1.28"
+    msg_in_129 = "deprecated in 1.29"
     msg_in_next = "deprecated in next version"
 
-    def deprecated_in_126(self) -> None:
-        """Emit a message about deprecation in 1.25."""
-        warnings.warn(self.msg_in_126, category=RemovedInDjangoCA126Warning)
-
     def deprecated_in_127(self) -> None:
-        """Emit a message about deprecation in 1.24."""
+        """Emit a message about deprecation in 1.27."""
         warnings.warn(self.msg_in_127, category=RemovedInDjangoCA127Warning)
+
+    def deprecated_in_128(self) -> None:
+        """Emit a message about deprecation in 1.28."""
+        warnings.warn(self.msg_in_128, category=RemovedInDjangoCA128Warning)
+
+    def deprecated_in_129(self) -> None:
+        """Emit a message about deprecation in 1.29."""
+        warnings.warn(self.msg_in_129, category=RemovedInDjangoCA129Warning)
 
     def deprecated_in_next(self) -> None:
         """Emit a message about deprecation in the next version."""
@@ -50,10 +56,12 @@ class TestDjangoCATestCase(TestCase):
     def test_base(self) -> None:
         """Test warning messages."""
 
-        with self.assertWarnsRegex(RemovedInDjangoCA126Warning, rf"^{self.msg_in_126}$"):
-            self.deprecated_in_126()
         with self.assertWarnsRegex(RemovedInDjangoCA127Warning, rf"^{self.msg_in_127}$"):
             self.deprecated_in_127()
+        with self.assertWarnsRegex(RemovedInDjangoCA128Warning, rf"^{self.msg_in_128}$"):
+            self.deprecated_in_128()
+        with self.assertWarnsRegex(RemovedInDjangoCA129Warning, rf"^{self.msg_in_129}$"):
+            self.deprecated_in_129()
         with self.assertWarnsRegex(RemovedInNextVersionWarning, rf"^{self.msg_in_next}$"):
             self.deprecated_in_next()